refactor: extract linewriter to package
